# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline AI DevOps Platform API
  version: 1.0.0
  description: |
    AI DevOps automation platform for deploying applications to Vercel and Render.
    Provides sandboxed workspace execution via Daytona, AI-powered repository analysis,
    and automated deployment orchestration.
  contact:
    name: Pipeline Support
    email: support@pipeline.dev
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.pipeline.dev/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /repos/connect:
    post:
      summary: Connect a Git repository
      description: Connect a Git repository from GitHub or GitLab to the platform
      tags: [Repositories]
      operationId: connectRepo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepoConnectionRequest'
      responses:
        '200':
          description: Repository connected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoConnection'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /repos/{repo_id}/analyze:
    post:
      summary: Analyze repository structure
      description: |
        Creates a Daytona workspace, clones the repository, and runs AI analysis
        to detect frameworks, services, and recommend deployment platforms.
      tags: [Repositories]
      operationId: analyzeRepo
      parameters:
        - name: repo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoAnalysis'
        '404':
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /repos/{repo_id}:
    get:
      summary: Get repository details
      tags: [Repositories]
      operationId: getRepo
      parameters:
        - name: repo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Repository details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoConnection'
        '404':
          description: Repository not found

  # ==================== DEPLOYMENTS ====================
...
  /sandboxes:
    get:
      summary: List all sandboxes
      tags: [Sandboxes]
      operationId: listSandboxes
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [creating, running, stopped, error, destroyed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of sandboxes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxList'

    post:
      summary: Create sandbox
      description: Create a new Daytona workspace for isolated execution
      tags: [Sandboxes]
      operationId: createSandbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SandboxCreateRequest'
      responses:
        '201':
          description: Sandbox created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
  /sandboxes/{sandbox_id}:
    get:
      summary: Get sandbox details
      tags: [Sandboxes]
      operationId: getSandbox
      parameters:
        - name: sandbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sandbox details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'

    delete:
      summary: Destroy sandbox
      description: Destroy the Daytona workspace and free resources
      tags: [Sandboxes]
      operationId: destroySandbox
      parameters:
        - name: sandbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Sandbox destroyed
  /sandboxes/{sandbox_id}/start:
    post:
      summary: Start sandbox
      tags: [Sandboxes]
      operationId: startSandbox
      parameters:
        - name: sandbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sandbox started
  /sandboxes/{sandbox_id}/stop:
    post:
      summary: Stop sandbox
      tags: [Sandboxes]
      operationId: stopSandbox
      parameters:
        - name: sandbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sandbox stopped
  /sandboxes/{sandbox_id}/execute:
    post:
      summary: Execute command in sandbox
      description: Run a command inside the Daytona workspace
      tags: [Sandboxes]
      operationId: executeCommand
      parameters:
        - name: sandbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandExecuteRequest'
      responses:
        '200':
          description: Command executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandExecuteResponse'
  /sandboxes/{sandbox_id}/terminal:
    post:
      summary: Open terminal session
      description: Open an interactive terminal session inside the sandbox
      tags: [Sandboxes]
      operationId: openTerminal
      parameters:
        - name: sandbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalSessionRequest'
      responses:
        '201':
          description: Terminal session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalSession'
  /sandboxes/{sandbox_id}/logs:
    get:
      summary: Get sandbox logs
      description: Retrieve execution logs from the sandbox
      tags: [Sandboxes]
      operationId: getSandboxLogs
      parameters:
        - name: sandbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tail
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Sandbox logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxLogs'

  # ==================== HEALTH ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RepoConnectionRequest:
      type: object
      required: [repo_url, provider]
      properties:
        repo_url:
          type: string
          format: uri
          description: Full URL to the Git repository
          example: https://github.com/user/my-monorepo
        provider:
          type: string
          enum: [github, gitlab]
          description: Git provider
          example: github
        branch:
          type: string
          description: Default branch to use
          default: main
          example: main
        name:
          type: string
          description: Optional custom name for the repository
          example: my-monorepo
    RepoConnection:
      type: object
      required: [id, repo_url, provider, status, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the connected repository
        repo_url:
          type: string
          format: uri
        provider:
          type: string
          enum: [github, gitlab]
        branch:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [pending, connected, analyzing, error]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DetectedService:
      type: object
      required: [name, framework, path]
      properties:
        name:
          type: string
          description: Service name (e.g., frontend, api, backend)
          example: frontend
        framework:
          type: string
          description: Detected framework
          example: nextjs
          enum: [nextjs, react, vue, angular, svelte, nuxtjs, remix, fastapi, flask, django, express, nestjs, go, rust, unknown]
        path:
          type: string
          description: Relative path within the repository
          example: frontend/
        language:
          type: string
          description: Primary programming language
          example: typescript
          enum: [javascript, typescript, python, go, rust, java, ruby, php, unknown]
        recommended_platform:
          type: string
          enum: [vercel, render, docker]
          description: Recommended deployment platform
        detected_files:
          type: array
          items:
            type: string
          description: Key files that helped identify the service
          example: [package.json, next.config.js]
        build_command:
          type: string
          description: Detected or suggested build command
          example: npm run build
        start_command:
          type: string
          description: Detected or suggested start command
          example: npm start
        env_variables:
          type: array
          items:
            type: string
          description: Detected environment variable requirements
    RepoAnalysis:
      type: object
      required: [repo_id, services, is_monorepo, analyzed_at]
      properties:
        repo_id:
          type: string
          format: uuid
        services:
          type: array
          items:
            $ref: '#/components/schemas/DetectedService'
        is_monorepo:
          type: boolean
          description: Whether the repository is a monorepo
        detected_workspaces:
          type: array
          items:
            type: string
          description: Workspace tools detected (e.g., turborepo, nx, pnpm)
        root_config:
          type: object
          description: Root-level configuration files
        analyzed_at:
          type: string
          format: date-time
        sandbox_id:
          type: string
          format: uuid
          description: ID of the sandbox used for analysis

    # ==================== DEPLOYMENT SCHEMAS ====================
...
    LogEntry:
      type: object
      required: [timestamp, level, message]
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [debug, info, warning, error, fatal]
        message:
          type: string
        source:
          type: string
          description: Component that generated the log

    # ==================== SANDBOX SCHEMAS ====================
    SandboxCreateRequest:
      type: object
      properties:
        repo_id:
          type: string
          format: uuid
          description: Optional repo to pre-clone
        repo_url:
          type: string
          format: uri
        branch:
          type: string
          default: main
        environment_variables:
          type: object
          additionalProperties:
            type: string
        resources:
          $ref: '#/components/schemas/SandboxResources'
    SandboxResources:
      type: object
      properties:
        cpu_cores:
          type: integer
          default: 2
        memory_mb:
          type: integer
          default: 4096
        disk_gb:
          type: integer
          default: 20
    Sandbox:
      type: object
      required: [id, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [creating, running, stopped, error, destroyed]
        repo_id:
          type: string
          format: uuid
        repo_url:
          type: string
          format: uri
        branch:
          type: string
        workspace_url:
          type: string
          description: URL to access the Daytona workspace
        resources:
          $ref: '#/components/schemas/SandboxResources'
        environment_variables:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        stopped_at:
          type: string
          format: date-time
        destroyed_at:
          type: string
          format: date-time
    SandboxList:
      type: object
      required: [sandboxes, total]
      properties:
        sandboxes:
          type: array
          items:
            $ref: '#/components/schemas/Sandbox'
        total:
          type: integer
    CommandExecuteRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
          example: npm install
        working_directory:
          type: string
          default: /workspace
        timeout_seconds:
          type: integer
          default: 300
        env_variables:
          type: object
          additionalProperties:
            type: string
    CommandExecuteResponse:
      type: object
      required: [exit_code, stdout, stderr, duration_ms]
      properties:
        exit_code:
          type: integer
        stdout:
          type: string
        stderr:
          type: string
        duration_ms:
          type: integer
        executed_at:
          type: string
          format: date-time
    TerminalSessionRequest:
      type: object
      properties:
        shell:
          type: string
          default: /bin/bash
        working_directory:
          type: string
          default: /workspace
        environment_variables:
          type: object
          additionalProperties:
            type: string
    TerminalSession:
      type: object
      required: [session_id, sandbox_id, websocket_url, created_at]
      properties:
        session_id:
          type: string
          format: uuid
        sandbox_id:
          type: string
          format: uuid
        websocket_url:
          type: string
          format: uri
          description: WebSocket URL for terminal connection
        shell:
          type: string
        working_directory:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    SandboxLogs:
      type: object
      required: [sandbox_id, logs]
      properties:
        sandbox_id:
          type: string
          format: uuid
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'

    # ==================== ERROR SCHEMAS ====================
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        request_id:
          type: string
          description: Request ID for debugging

    # ==================== HEALTH SCHEMAS ====================
...